+++ CHANGELOG.md
@@ -6,9 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- LinearizedPredictor
+- SummedPredictor
+- ConcatenatedPredictor
+- LinearEinsumPredictor
+- Support for "same" systematics in ComposedPredictors
+
 ### Changed
 - Calculate M. distance truth-bin by bin to avoid memory issues.
 - Relax requirements for exact Matplotlib version.
+- Renamed LinearPredictor to MatrixPredictor
 
 ### Removed
 - Removed support for Python versions below 3.7, test up to 3.11 now.